Implemented non-integer powers and toughness'
Added artist and flavor search

Pressing the search key on the search page starts the search

Fixed character bug in MtgXml2Json

+++ trunk/MTGFamiliar/src/com/mtg/fam/MtgCard.java
@@ -7,8 +7,8 @@ public class MtgCard {
 	public char rarity;
 	public String manacost;
 	public int cmc;
-	public int power;
-	public int toughness;
+	public float power;
+	public float toughness;
 	public int loyalty;
 	public String ability;
 	public String flavor;
